backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-docx==0.8.11
lxml==4.9.3
supabase==2.0.3
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
openai==1.3.9
google-generativeai==0.3.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
typing-extensions==4.8.0
